Update tests to add support for custom runner
I've added support for the custom runner to the tests. The main change
is that the names of the classes and test methods now include characters
that force the correct order. (Normally, unittests are run in "random"
order as they should be independent.)

Other changes include ensuring tests fail instead producing an error if
a challenger has yet to implement something.